Namespace MiqRequestTaskHelper to Spec::Support::MiqRequestTaskHelper
