+++ spec/models/miq_provision/state_machine_spec_helper.rb
@@ -1,5 +1,5 @@
 module MiqProvision::StateMachineSpecHelper
-  include MiqRequestTaskHelper
+  include Spec::Support::MiqRequestTaskHelper
 
   def test_prepare_provision
     call_method
